Reuse doxygen docstrings in our rst file (for procCcd)
This is how we can use the API comments currently in the doxygen
docstrings and incorporate them into our own rst file, using a mockup
written by JSick.  It duplicates other parts of the page (e.g. the
Entrypoint), currently, but has a new section on Initzn, and is a good
first draft of how we might go about this.

(For conf.py: during sphinx processing there was a warning that
version wanted a string not a float, so i changed that to no longer
get the warning.)