Fill in examples section with code extracts
We want to provide some of the most relevant lines of the example
script code so that users know what's happening when they are running
it, and we so far do this along the lines of the choices the doxygen
string writers did, selecting those same parts.

Also modified links in this section so they point to relevant places,
like the current EUPS guide.