Insert code lines from the runIsrTask example
This brings back a fair amount of what i'd written earlier into the
example section to fill it out and detail what's going on with
runIsrTask more thoroughly.  The core is from the doxygen pages, but i
added more detail and interstitial explanation also, which i think is
something useful we can spell out for developers.