Reduce double backticked items + insert links
For these 2 Tasks so far, have moved most of the double backticked
items into single backticked ones, and then inserted relevant links to
doxygen and github pages wherever relevant and easily findable.  A few
links are placeholders, not going anywhere currently, but planned to
go to a page we make in the future, or probably there already exists
one that hasn't been easily found yet.