Change double to single backticks for charimg
Now for charimg: have moved most of the double backticked items into
single backticked ones, and again inserted relevant links to doxygen
and github pages wherever relevant and easily findable.  A few links
are placeholders, not going anywhere currently, but planned to go to a
page we make in the future, or probably there already exists one that
hasn't been easily found yet.