Insert rst comments on links
This is a wholesale entering of multiple rst comments in all sfp files
explaining that we'll want to link to pages for many of the technical
catalog/exposure etc. types.  These aren't given in much detail or
precision other than to say we need them for now, we certainly want to
figure out the framework for these pages.

In other places, comments are given to why the particular doxygen page
was chosen to be linked to.

(In some cases the comments are a few lines down from the ref because
Sphinx gives warnings about lines not being ended properly if i
attempt to interleave them in a list.)