Insert Confluence page with i/o's into isrtask
Inserted Hsin-Fang's page which has useful very concise info into this
specific page, because isrtask has so many inputs, though it could be
somewhat useful for all the sfp pages.  Not sure we should be linking
to a page like this, but in the interest of more info being
potentially useful, i did so for now.