Make sure all default params are code literals
This is now done in all 4 sfp codes, so that everything should render
as a code literal, *except* for 'no default', which i've simply left
as is for now (this is only in isrTask).  I also changed these vars in
the body of the table to code literals where they appear.