Update Task Template with sfp Task learnings
This is the first full draft of what the Task Template will look like
as handed off to DM developers.  It contains many modifications and
much more content from we have learned while writing out the 4 sfp
Tasks in detail by hand.  It's written largely in the imperative form
as seemed appropriate when passing on, and further delineates the
sections that will be filled in automatically, and those that will
need docstrings actually written (which is the majority of it).  There
are a number of comments in the rst in this version for places that
there is not yet full clarity on what's to be done.