Clarify content for strings vs. docstrings
This is primarily in the Guidance doc, where comparing with the
doxygen placement of docstrings, i now refer more precisely to what
goes in which part of the strings of the code -- which will be picked
up automatically anyway, so devs won't need to do anything 'extra'
here, and that should be stated.  But at least, it's clear now what
should be said about the string parts in the guidance doc.